requirements.txt
beautifulsoup4==4.11.1
bleach==4.1.0
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
click==8.1.3
cmake==3.27.2
coloredlogs==15.0.1
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
Django==4.2.1
django-progressbarupload==0.1.7
Flask==2.2.3
Flask-Cors==3.0.10
grpcio==1.40.0
hiq-python==1.1.12
huggingface-hub==0.20.1
invisible-watermark==0.1.5
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-language-pack-zh-CN @ http://autodl-public.ks3-cn-beijing.ksyun.com/instance/jupyterlab_language_pack_zh_CN-0.0.1.dev0-py2.py3-none-any.whl
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.8.0
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
moviepy==1.0.3
networkx==2.8.8
notebook==6.4.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.1
onnx==1.15.0
onnxruntime==1.16.2
opencv-python==4.6.0.66
packaging==21.0
pandas==1.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pyllama==0.0.3
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
safetensors==0.3.3
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.11.2
sklearn==0.0.post1
soupsieve==2.3.2.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==2.0.1
torchvision==0.15.2
tqdm @ file:///root/autodl-tmp/yolov5/tqdm-4.64.0-py2.py3-none-any.whl
transformers==4.19.2
typing-extensions==4.8.0
ultralytics==8.0.199
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
websocket-client==1.2.1
zipp==3.14.0
